Nutrinfo

Documentation for Nutrinfo.

Nutrinfo._add_components

Nutrinfo._find_nv_in_db

Nutrinfo.add_edge

Nutrinfo.add_vertex

Nutrinfo.list_cycles

Nutrinfo.list_edges

Nutrinfo.list_frontier

Nutrinfo.list_paths

Nutrinfo.make_graph

Nutrinfo.reduce_path

Nutrinfo.resolve

Nutrinfo._add_components — Method

Nutrinfo.resolve helper function.

Adds all the Components of the supplied NutrientVector src as edges from the existing vertex for src to either new or existing vertices for the Components of src.

Nutrinfo._find_nv_in_db — Method

Nutrinfo.resolve helper function.

Try to find the NutrientVector with name name in the supplied NutrientVector database nvDB, and return the collection of matching NutrientVectors.

If this collection is empty, there is no match, and the algorithm regards the NutrientVector in question as a basis vector. (It could also mean that the database is missing an entry that should be there, and it is up to the user to decide whether that is the case, and to add it if possible.)

If the collection has exactly one member, that is a match.

If it has more than one member, then it is flagged as an error by the algorithm, and the database needs to be fixed to resolve the duplication.

Nutrinfo.add_edge — Method

Nutrinfo.resolve helper function.

Adds an edge to our MetaDiGraph representing the relationship that NutrientVector src has quantity qty of NutrientVector dst as a Component.

Nutrinfo.add_vertex — Method

Nutrinfo.resolve helper function.

Adds a vertex to our MetaDiGraph representing NutrientVector nv.

Nutrinfo.list_cycles — Method

Nutrinfo.resolve helper function.

Lists all the cycles in our MetaDiGraph. We require that there be no cycles if we are going to be able to use list_paths(). So this function should be called before calling list_paths() to ensure that the latter terminates.

Nutrinfo.list_edges — Method

Nutrinfo.resolve helper function.

Lists all edges in our MetaDiGraph as three element arrays of the form [src,qty,dst], representing the relationship that NutrientVector src has quantity qty of NutrientVector dst as a Component.

Nutrinfo.list_frontier — Method

Nutrinfo.resolve helper function.

Lists all vertices in our MetaDiGraph with outdegree zero. These correspond to NutrientVectors that either are terminal basis NutrientVectors, or are NutrientVectors that we haven't yet tried to expand.

Nutrinfo.list_paths — Method

Nutrinfo.resolve helper function.

List all possible paths in the supplied MetaDiGraph g that start at vertex src.

Each path is an odd-length array of strictly alternating names of vertices and edges, beginning with the vertex src.

This can be expected not to terminate if g contains any cycles, so use list_cycles() to determine that first.

Nutrinfo.make_graph — Method

Nutrinfo.resolve helper function.

Creates a MetaDiGraph where vertices represent NutrientVectors, and directed edges represent the Component relationships between them.

Nutrinfo.reduce_path — Method

Nutrinfo.resolve helper function.

Take a path (produced by list_paths()) and "reduce" it to the equivalent "weight" of Component denoted by its terminal vertex.

Multiply all the edges together, and then divide by all the servings of internal vertices

edges are path[2], path[4], ..., path[length(path)-1]

internal vertices are path[3], path[5], ..., path[length(path)-2]

The point of this exercise is to create Components for all paths returned by list_paths() so that they can be added together to get the Components of the output of resolve().

Nutrinfo.resolve — Method

The "resolve" NutrientVector operation

Nutrinfo.jl is a program for tracking energy and macronutrient consumption.

It takes as input two JSON files:

A nutrient information database, which contains nutritional information about a collection of foodstuffs as an array of JSON objects. This JSON array is validated using https://github.com/fredo-dedup/JSONSchema.jl to ensure that the nutritional information fields are recognised by the program.

A food diary, which contains information about the user's daily intake of foodstuffs over a period of time.

For now, it produces as output a daily summary of the total energy, macronutrient quantity and macronutrient ratio consumed.

(In future it could be extended to supply micronutrient quantity information and create plots.)

It makes use of https://github.com/JuliaIntervals/IntervalArithmetic.jl to deal with uncertainty in information about nutritional constituents and/or quantities of consumed foodstuffs.

It makes use of https://github.com/PainterQubits/Unitful.jl to deal with standard (energy, capacity, mass) as well as ad-hoc (approximate serving sizes) needs for appropriately processing the kinds of quantities we'll be dealing with.